requirements.txt
openai>=1.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
docling>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
